refactor(shared-terminal): split feedback/__init__.py into named modules
Splits the single-file surfaces/shared/terminal/feedback/__init__.py into
persistence.py, analytics.py, context_display.py, prompts.py,
miss_classification.py, and flow.py, each owning one concern. __init__.py
now only re-exports the public prompt_investigation_feedback entrypoint.

Renames the ANSI escape constants in prompts.py from the terse _H/_D/_R to
_HIGHLIGHT/_DIM/_RESET for readability; all other private helper names are
preserved across the new module boundaries per the issue's guidance, since
tests reach into them directly.

Closes #5444
